Feat: 구독자 급등 작가 클릭시 작가 홈으로 이동 기능 추가 Modify: 코드 경로 수정 Design: 작가 div에 cursor 효과 추가 Rename: src/sign-up 폴더 구조를 src/user/img, login, sign-up 구조로 변경 Remove: src/main.ts, style.css 삭제
