+++ src/main-page/main-page.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./main-page.css" />
+    <link rel="stylesheet" href="../main-page/main-page.css" />
     <link rel="stylesheet" href="../common/index.css" />
     <script type="module" src="../components/header.ts"></script>
     <script type="module" src="../components/footer.ts"></script>
@@ -12,58 +12,5 @@
   </head>
   <body>
     <main-page></main-page>
-    <!-- 원래 main-page html 내용 -->
-    <!-- <div class="main-page">
-      <div class="banner">
-        <p class="book-title">멈추어 버린 시간</p>
-        <p class="book-writer">by 김수정</p>
-        <img src="./img/banner.png" alt="배너 사진" />
-      </div>
-      <div class="brunch">
-        <div class="brunch-list-title">
-          <p>Today's Pick</p>
-          <h2>요즘 뜨는 브런치</h2>
-        </div>
-        <ul class="brunch-list"></ul>
-      </div>
-
-      <div class="hot-writer">
-        <h3>구독자 급등 작가</h3>
-        <div class="writer-list">
-          <div class="writer1">
-            <figure class="writer">
-              <img src="./img/writer1.png" alt="" width="" height="" />
-              <figcaption class="title">피터팬</figcaption>
-              <figcaption class="job">목사 만화가</figcaption>
-              <figcaption class="text">하나님을 사랑하고 아내를 사랑하고 여행을 ...</figcaption>
-            </figure>
-          </div>
-          <div class="writer2">
-            <figure class="writer">
-              <img src="./img/writer2.png" alt="" width="" height="" />
-              <figcaption class="title">빛날현</figcaption>
-              <figcaption class="job">에세이스트</figcaption>
-              <figcaption class="text">26살부터 47살까지 교수님이라 불리우는 목...</figcaption>
-            </figure>
-          </div>
-          <div class="writer3">
-            <figure class="writer">
-              <img src="./img/writer3.png" alt="" width="" height="" />
-              <figcaption class="title">허사이</figcaption>
-              <figcaption class="job"></figcaption>
-              <figcaption class="text">책방 셔터맨을 꿈꾸는 40대 공돌이. 다양한 ...</figcaption>
-            </figure>
-          </div>
-          <div class="writer4">
-            <figure class="writer">
-              <img src="./img/writer4.png" alt="" width="" height="" />
-              <figcaption class="title">마틸다 하나씨</figcaption>
-              <figcaption class="job">에세이스트</figcaption>
-              <figcaption class="text">하노이에 깃든 시간을 따라 매일 새로운 세...</figcaption>
-            </figure>
-          </div>
-        </div>
-      </div>
-    </div> -->
   </body>
 </html>
