+++ src/user/login/login.html
@@ -3,15 +3,16 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="../sign-up/login.css" />
+    <link rel="stylesheet" href="../common/index.css" />
+    <link rel="stylesheet" href="../login/login.css" />
     <!-- import시 type="module"를 작성해서 ts파일을 불러올 수 있다. -->
-    <script type="module" defer src="../sign-up/login.ts"></script>
+    <script type="module" defer src="../login/login.ts"></script>
     <title>Login</title>
   </head>
   <body>
     <div class="login">
       <div class="login-banner">
-        <img src="../sign-up/img/login.png" alt="책 읽는 모습" />
+        <img src="../img/login.png" alt="책 읽는 모습" />
         <h3>브런치 작가로 데뷔하세요.</h3>
         <p class="text">
           진솔한 에세이부터 업계 전문 지식까지,<br />
