+++ src/user/sign-up/sign-up.html
@@ -3,8 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script type="module" defer src="../sign-up/sign-up.ts"></script>
+    <link rel="stylesheet" href="../common/index.css" />
     <link rel="stylesheet" href="../sign-up/sign-up.css" />
+    <script type="module" defer src="../sign-up/sign-up.ts"></script>
     <title>SignUp</title>
   </head>
   <body>
